requirements.txt
streamlit==1.31.0
groq
python-dotenv==1.0.0
